objprint: int count in objFindTexture (90.8->98.8%)
Typing the textureSlotCount local as int instead of u8 defers the
count load to just before the loop (matching the target's lbz at the
modelDef base) and lets the entries pointer recycle the modelDef
register, recovering the load order and register reuse.

+++ src/main/objprint.c
@@ -1328,7 +1328,7 @@ ObjTextureRuntimeSlot* objFindTexture(void* obj, int target, int unusedMaterialI
         ObjTextureSlotDef* entries = modelDef->textureSlotDefs;
         if (entries == NULL) return NULL;
         {
-            u8 count = modelDef->textureSlotCount;
+            int count = modelDef->textureSlotCount;
             int i;
             for (i = 0; i < count; i++)
             {
